azm: Move the AMD64 bits into its own namespace
We'll want to be able to assemble e.g. aarch64 too in the future.
